Split 5 oversized files to comply with 1000-line limit
- tab_bar/mod.rs (1216->820): Extract view rendering to render.rs,
  Tab impl methods to tab.rs, truncate_label helper to render.rs
- table/mod.rs (1036->938): Extract view rendering to render.rs
- loading_list/mod.rs (1032->959): Extract view rendering to render.rs
- chart/mod.rs (1017->875): Extract DataSeries impl methods to series.rs
- integration_new_components.rs (1834->930+959): Split tests 1-14 and
  15-30 into two files, clean up unused imports in each
