+++ src/component/tab_bar/mod.rs
@@ -32,12 +32,14 @@
 //! ```
 
 use ratatui::prelude::*;
-use ratatui::widgets::Paragraph;
-use unicode_width::UnicodeWidthStr;
 
 use super::{Component, Disableable, Focusable};
 use crate::input::{Event, KeyCode};
 use crate::theme::Theme;
+
+mod render;
+#[cfg(test)]
+use render::truncate_label;
 
 // ---------------------------------------------------------------------------
 // Tab
@@ -70,250 +72,19 @@ use crate::theme::Theme;
 )]
 pub struct Tab {
     /// Unique identifier for this tab.
-    id: String,
+    pub(super) id: String,
     /// Display label.
-    label: String,
+    pub(super) label: String,
     /// Whether the tab can be closed.
-    closable: bool,
+    pub(super) closable: bool,
     /// Whether the tab has unsaved modifications.
-    modified: bool,
+    pub(super) modified: bool,
     /// Optional icon prefix (e.g., a file-type indicator).
-    icon: Option<String>,
+    pub(super) icon: Option<String>,
 }
 
-impl Tab {
-    /// Creates a new tab with the given id and label.
-    ///
-    /// By default the tab is not closable, not modified, and has no icon.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use envision::component::Tab;
-    ///
-    /// let tab = Tab::new("t1", "Overview");
-    /// assert_eq!(tab.id(), "t1");
-    /// assert_eq!(tab.label(), "Overview");
-    /// assert!(!tab.closable());
-    /// assert!(!tab.modified());
-    /// assert_eq!(tab.icon(), None);
-    /// ```
-    pub fn new(id: impl Into<String>, label: impl Into<String>) -> Self {
-        Self {
-            id: id.into(),
-            label: label.into(),
-            closable: false,
-            modified: false,
-            icon: None,
-        }
-    }
-
-    /// Sets whether the tab is closable (builder).
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use envision::component::Tab;
-    ///
-    /// let tab = Tab::new("t1", "File").with_closable(true);
-    /// assert!(tab.closable());
-    /// ```
-    pub fn with_closable(mut self, closable: bool) -> Self {
-        self.closable = closable;
-        self
-    }
-
-    /// Sets whether the tab is modified (builder).
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use envision::component::Tab;
-    ///
-    /// let tab = Tab::new("t1", "File").with_modified(true);
-    /// assert!(tab.modified());
-    /// ```
-    pub fn with_modified(mut self, modified: bool) -> Self {
-        self.modified = modified;
-        self
-    }
-
-    /// Sets an icon for the tab (builder).
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use envision::component::Tab;
-    ///
-    /// let tab = Tab::new("t1", "File").with_icon("R");
-    /// assert_eq!(tab.icon(), Some("R"));
-    /// ```
-    pub fn with_icon(mut self, icon: impl Into<String>) -> Self {
-        self.icon = Some(icon.into());
-        self
-    }
-
-    /// Returns the tab's unique identifier.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use envision::component::Tab;
-    ///
-    /// let tab = Tab::new("editor-1", "main.rs");
-    /// assert_eq!(tab.id(), "editor-1");
-    /// ```
-    pub fn id(&self) -> &str {
-        &self.id
-    }
-
-    /// Returns the tab's display label.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use envision::component::Tab;
-    ///
-    /// let tab = Tab::new("t1", "Overview");
-    /// assert_eq!(tab.label(), "Overview");
-    /// ```
-    pub fn label(&self) -> &str {
-        &self.label
-    }
-
-    /// Returns whether the tab is closable.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use envision::component::Tab;
-    ///
-    /// let tab = Tab::new("t1", "File");
-    /// assert!(!tab.closable());
-    ///
-    /// let tab = Tab::new("t1", "File").with_closable(true);
-    /// assert!(tab.closable());
-    /// ```
-    pub fn closable(&self) -> bool {
-        self.closable
-    }
-
-    /// Returns whether the tab has unsaved modifications.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use envision::component::Tab;
-    ///
-    /// let tab = Tab::new("t1", "File").with_modified(true);
-    /// assert!(tab.modified());
-    /// ```
-    pub fn modified(&self) -> bool {
-        self.modified
-    }
-
-    /// Returns the tab's icon, if any.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use envision::component::Tab;
-    ///
-    /// let tab = Tab::new("t1", "File");
-    /// assert_eq!(tab.icon(), None);
-    ///
-    /// let tab = Tab::new("t1", "File").with_icon("R");
-    /// assert_eq!(tab.icon(), Some("R"));
-    /// ```
-    pub fn icon(&self) -> Option<&str> {
-        self.icon.as_deref()
-    }
-
-    /// Sets the label.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use envision::component::Tab;
-    ///
-    /// let mut tab = Tab::new("t1", "Old Name");
-    /// tab.set_label("New Name");
-    /// assert_eq!(tab.label(), "New Name");
-    /// ```
-    pub fn set_label(&mut self, label: impl Into<String>) {
-        self.label = label.into();
-    }
-
-    /// Sets the closable flag.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use envision::component::Tab;
-    ///
-    /// let mut tab = Tab::new("t1", "File");
-    /// tab.set_closable(true);
-    /// assert!(tab.closable());
-    /// ```
-    pub fn set_closable(&mut self, closable: bool) {
-        self.closable = closable;
-    }
-
-    /// Sets the modified flag.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use envision::component::Tab;
-    ///
-    /// let mut tab = Tab::new("t1", "File");
-    /// tab.set_modified(true);
-    /// assert!(tab.modified());
-    /// ```
-    pub fn set_modified(&mut self, modified: bool) {
-        self.modified = modified;
-    }
-
-    /// Sets the icon.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use envision::component::Tab;
-    ///
-    /// let mut tab = Tab::new("t1", "File");
-    /// tab.set_icon(Some("R".to_string()));
-    /// assert_eq!(tab.icon(), Some("R"));
-    ///
-    /// tab.set_icon(None);
-    /// assert_eq!(tab.icon(), None);
-    /// ```
-    pub fn set_icon(&mut self, icon: Option<String>) {
-        self.icon = icon;
-    }
-
-    /// Returns the rendered width of this tab, including decorations.
-    ///
-    /// Layout: ` [icon ]label[modified][close] `
-    fn rendered_width(&self, max_tab_width: Option<usize>) -> usize {
-        let mut w: usize = 2; // leading and trailing space
-        if let Some(icon) = &self.icon {
-            w += icon.width() + 1; // icon + space
-        }
-        w += self.label.width();
-        if self.modified {
-            w += 1; // bullet
-        }
-        if self.closable {
-            w += 2; // space + close char
-        }
-        if let Some(max) = max_tab_width {
-            w.min(max)
-        } else {
-            w
-        }
-    }
-}
+// Tab methods are in tab.rs
+mod tab;
 
 // ---------------------------------------------------------------------------
 // TabBarMessage
@@ -1021,150 +792,7 @@ impl Component for TabBar {
     }
 
     fn view(state: &Self::State, frame: &mut Frame, area: Rect, theme: &Theme) {
-        if area.height == 0 || area.width == 0 {
-            return;
-        }
-
-        let available_width = area.width as usize;
-
-        // Build rendered spans for each tab, computing widths.
-        struct RenderedTab {
-            spans: Vec<Span<'static>>,
-            width: usize,
-        }
-
-        let rendered: Vec<RenderedTab> = state
-            .tabs
-            .iter()
-            .enumerate()
-            .map(|(i, tab)| {
-                let is_active = state.active == Some(i);
-                let base_style = if state.disabled {
-                    theme.disabled_style()
-                } else if is_active {
-                    theme.focused_style().add_modifier(Modifier::BOLD)
-                } else {
-                    theme.normal_style()
-                };
-
-                let mut parts: Vec<Span<'static>> = Vec::new();
-                parts.push(Span::styled(" ", base_style));
-
-                if let Some(icon) = &tab.icon {
-                    parts.push(Span::styled(format!("{icon} "), base_style));
-                }
-
-                // Label (possibly truncated)
-                let label_text = if let Some(max) = state.max_tab_width {
-                    let decoration_width = tab.rendered_width(None) - tab.label.width();
-                    let max_label = max.saturating_sub(decoration_width);
-                    truncate_label(&tab.label, max_label)
-                } else {
-                    tab.label.clone()
-                };
-                parts.push(Span::styled(label_text, base_style));
-
-                if tab.modified {
-                    let mod_style = if state.disabled {
-                        theme.disabled_style()
-                    } else {
-                        theme.warning_style()
-                    };
-                    parts.push(Span::styled("*", mod_style));
-                }
-
-                if tab.closable {
-                    let close_style = if state.disabled {
-                        theme.disabled_style()
-                    } else {
-                        theme.error_style()
-                    };
-                    parts.push(Span::styled(" x", close_style));
-                }
-
-                parts.push(Span::styled(" ", base_style));
-
-                let width = tab.rendered_width(state.max_tab_width);
-                RenderedTab {
-                    spans: parts,
-                    width,
-                }
-            })
-            .collect();
-
-        // Determine which tabs are visible starting from scroll_offset.
-        let has_left_overflow = state.scroll_offset > 0;
-        let indicator_width: usize = 2; // "< " or " >"
-
-        let usable_left = if has_left_overflow {
-            available_width.saturating_sub(indicator_width)
-        } else {
-            available_width
-        };
-
-        // Walk from scroll_offset to find how many tabs fit.
-        let mut used = 0usize;
-        let mut visible_end = state.scroll_offset;
-        for rt in rendered.iter().skip(state.scroll_offset) {
-            let needed = used + rt.width;
-            if needed > usable_left {
-                break;
-            }
-            used = needed;
-            visible_end += 1;
-        }
-
-        // Check right overflow; if there are more tabs, reserve indicator space.
-        let has_right_overflow = visible_end < rendered.len();
-        if has_right_overflow && visible_end > state.scroll_offset {
-            // Re-check if the last visible tab still fits with the indicator.
-            let total_with_indicator = used + indicator_width;
-            if total_with_indicator > available_width {
-                // Drop the last visible tab.
-                visible_end -= 1;
-            }
-        }
-
-        // Build the final Line.
-        let mut spans: Vec<Span<'static>> = Vec::new();
-
-        if has_left_overflow {
-            let indicator_style = if state.disabled {
-                theme.disabled_style()
-            } else {
-                theme.info_style()
-            };
-            spans.push(Span::styled("< ", indicator_style));
-        }
-
-        for rt in rendered
-            .iter()
-            .skip(state.scroll_offset)
-            .take(visible_end.saturating_sub(state.scroll_offset))
-        {
-            spans.extend(rt.spans.iter().cloned());
-        }
-
-        if has_right_overflow {
-            let indicator_style = if state.disabled {
-                theme.disabled_style()
-            } else {
-                theme.info_style()
-            };
-            spans.push(Span::styled(" >", indicator_style));
-        }
-
-        let line = Line::from(spans);
-        let paragraph = Paragraph::new(line);
-
-        let annotation = crate::annotation::Annotation::new(crate::annotation::WidgetType::TabBar)
-            .with_id("tab_bar")
-            .with_focus(state.focused)
-            .with_disabled(state.disabled)
-            .with_selected(state.active.is_some())
-            .with_value(state.active.map(|i| i.to_string()).unwrap_or_default());
-        let annotated = crate::annotation::Annotate::new(paragraph, annotation);
-        frame.render_widget(annotated, area);
+        render::render_tab_bar(state, frame, area, theme);
     }
 }
 
@@ -1186,30 +814,6 @@ impl Disableable for TabBar {
     fn set_disabled(state: &mut Self::State, disabled: bool) {
         state.disabled = disabled;
     }
-}
-
-/// Truncates a label to at most `max_width` display columns, appending
-/// an ellipsis if truncation occurs.
-fn truncate_label(label: &str, max_width: usize) -> String {
-    if label.width() <= max_width {
-        return label.to_string();
-    }
-    if max_width == 0 {
-        return String::new();
-    }
-    let mut result = String::new();
-    let mut w = 0;
-    let target = max_width.saturating_sub(1); // reserve 1 for ellipsis
-    for ch in label.chars() {
-        let cw = unicode_width::UnicodeWidthChar::width(ch).unwrap_or(0);
-        if w + cw > target {
-            break;
-        }
-        result.push(ch);
-        w += cw;
-    }
-    result.push('\u{2026}'); // ellipsis
-    result
 }
 
 #[cfg(test)]
